master Horses slowed acording to parameters

+++ backend/database/migrations/2019_08_31_183332_create_horses_table.php
@@ -21,7 +21,7 @@ class CreateHorsesTable extends Migration
 			$table->integer('endurance');
 			$table->string('status');
 			$table->integer('position');
-			$table->integer('distance_covered');
+			$table->float('distance_covered');
 			$table->timestamps();
 		});
 	}
